added paintpanel class, called it in main
Basically, it creates a paint canvas, 7 color rectangles, and a clear
button.  You can select the colors, but you can't paint anything as of
now.  Spacing for the buttons need to be adjusted for the future.  Also,
I think it currently layers over the ellipses.  Just putting what I've
got so far